+++ hc/api/management/commands/smtpd.py
@@ -14,17 +14,19 @@ from asgiref.sync import sync_to_async
 from django.conf import settings
 from django.core.management.base import BaseCommand
 from django.db import connection
-
 from hc.api.models import Check
 from hc.lib.html import html2text
 
 RE_UUID = re.compile(
-    "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-4[a-fA-F0-9]{3}-[8|9|aA|bB][a-fA-F0-9]{3}-[a-fA-F0-9]{12}$"
+    r"^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-4[a-fA-F0-9]{3}-[8|9|aA|bB][a-fA-F0-9]{3}-[a-fA-F0-9]{12}$"
 )
+
+RE_PING_KEY_SLUG = re.compile(r"^[a-zA-Z0-9_-]{22}\+[a-z0-9-_]+$")
 
 
 class LogSink(Protocol):
-    def write(self, msg: str) -> None: ...
+    def write(self, msg: str) -> None:
+        ...
 
 
 def _match(subject: str, keywords: str) -> bool:
@@ -56,21 +58,29 @@ def _to_text(message: EmailMessage, with_subject: bool, with_body: bool) -> str:
 
 
 def _process_message(remote_addr: str, mailfrom: str, mailto: str, data: bytes) -> str:
-    to_parts = mailto.split("@")
-    code = to_parts[0]
-
-    if not RE_UUID.match(code):
-        return f"Not an UUID: {code}"
-
     # Get a new db connection in case the old one has timed out.
     # The if condition makes sure this does not run during tests.
     if not connection.in_atomic_block:
         connection.close()
 
-    try:
-        check = Check.objects.get(code=code)
-    except Check.DoesNotExist:
-        return f"Check not found: {code}"
+    to_parts = mailto.split("@")
+    mbox = to_parts[0]
+    if "+" in mbox:
+        # Pinging by slug
+        ping_key, slug = mbox.split("+")
+        try:
+            check = Check.objects.get(slug=slug, project__ping_key=ping_key)
+        except Check.DoesNotExist:
+            return f"Check not found: {mailto}"
+        except Check.MultipleObjectsReturned:
+            return f"Ambiguous slug: {mailto}"
+    else:
+        # Pinging by code
+        code = mbox
+        try:
+            check = Check.objects.get(code=code)
+        except Check.DoesNotExist:
+            return f"Check not found: {mailto}"
 
     action = "success"
     if check.filter_subject or check.filter_body:
@@ -88,10 +98,10 @@ def _process_message(remote_addr: str, mailfrom: str, mailto: str, data: bytes) 
         elif check.start_kw and _match(text, check.start_kw):
             action = "start"
 
-    ua = "Email from %s" % mailfrom
+    ua = f"Email from {mailfrom}"
     check.ping(remote_addr, "email", "", ua, data, action, None)
 
-    return f"Processed ping for {code}"
+    return f"Processed ping for {mailto}"
 
 
 class PingHandler:
@@ -107,10 +117,10 @@ class PingHandler:
         address: str,
         rcpt_options: list[str],
     ) -> str:
-        code, domain = address.split("@", maxsplit=1)
+        mbox, domain = address.split("@", maxsplit=1)
         if domain != settings.PING_EMAIL_DOMAIN:
             return "550 5.1.1 Recipient rejected"
-        if not RE_UUID.match(code):
+        if not RE_UUID.match(mbox) and not RE_PING_KEY_SLUG.match(mbox):
             return "550 5.1.1 Invalid mailbox"
 
         envelope.rcpt_tos.append(address)