+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - New integration: Google Chat (#423)
 - Fix the ping view to handle IPv4-mapped IPv6 addresses (#1169)
 - Update Dockerfile to listen on IPv6 as well (#1177)
+- Add "ping_key+slug@..." address support when pinging by email (#1009)
 
 ## v3.10 - 2025-05-09